+++ sdk-java/src/main/java/ly/count/sdk/java/internal/InternalConfig.java
@@ -213,7 +213,7 @@ public class InternalConfig extends Config {
         return !locationEnabled;
     }
 
-    boolean isAutoSendUserPropertiesOnSessions() {
-        return autoSendUserPropertiesOnSessions;
+    boolean isAutoSendUserProperties() {
+        return autoSendUserProperties;
     }
 }